Version 4.2

+++ src/win_init.c
@@ -30,7 +30,7 @@ void	init_vars(t_game *g, t_sett *sett)
 	g->mlx = mlx_init();
 	get_img(g, &g->img, 0, 0);
 	g->pos = pos(g->sett->map);
-	g->cht = g->sett->map[g->pos.y][g->pos.x];
+	g->cht = g->sett->map[(int)g->pos.y][(int)g->pos.x];
 	g->height = matrix_len(g->sett->map) * g->img.width + 3;
 	g->width = max_strlen(g->sett->map) * g->img.height + 3;
 }
